+++ user-service/src/main/java/rs/raf/user_service/controller/EmployeeController.java
@@ -1,0 +1,67 @@
+package rs.raf.user_service.controller;
+
+import io.swagger.v3.oas.annotations.Operation;
+import io.swagger.v3.oas.annotations.responses.ApiResponse;
+import io.swagger.v3.oas.annotations.responses.ApiResponses;
+import org.springframework.data.domain.Page;
+import org.springframework.data.domain.PageRequest;
+import org.springframework.data.domain.Pageable;
+import org.springframework.web.bind.annotation.*;
+import rs.raf.user_service.entity.EmployeeDTO;
+import rs.raf.user_service.service.EmployeeService;
+
+@RestController
+@RequestMapping("/api/admin/employees")
+public class EmployeeController {
+
+    private final EmployeeService employeeService;
+
+    public EmployeeController(EmployeeService employeeService) {
+        this.employeeService = employeeService;
+    }
+
+    /*
+        @GetMapping("/{id}")
+        public Employee getEmployeeById(@PathVariable Long id) {
+            return employeeService.findById(id);
+        }
+
+        @GetMapping
+        public Page<Employee> getAllEmployees(
+                @RequestParam(required = false) String position,
+                @RequestParam(required = false) String department,
+                @RequestParam(required = false) Boolean active,
+                @RequestParam(defaultValue = "0") int page,
+                @RequestParam(defaultValue = "10") int size) {
+
+            Pageable pageable = PageRequest.of(page, size);
+            return employeeService.findAll(position, department, active, pageable);
+        }
+
+     */
+    @Operation(summary = "Get employee by ID", description = "Returns an employee based on the provided ID")
+    @ApiResponses(value = {
+            @ApiResponse(responseCode = "200", description = "Successfully retrieved employee"),
+            @ApiResponse(responseCode = "404", description = "Employee not found")
+    })
+    @GetMapping("/{id}")
+    public EmployeeDTO getEmployeeById(@PathVariable Long id) {
+        return employeeService.findById(id);
+    }
+
+    @Operation(summary = "Get all employees", description = "Returns a paginated list of employees with optional filters")
+    @ApiResponses(value = {
+            @ApiResponse(responseCode = "200", description = "Successfully retrieved employee list")
+    })
+    @GetMapping
+    public Page<EmployeeDTO> getAllEmployees(
+            @RequestParam(required = false) String position,
+            @RequestParam(required = false) String department,
+            @RequestParam(required = false) Boolean active,
+            @RequestParam(defaultValue = "0") int page,
+            @RequestParam(defaultValue = "10") int size) {
+
+        Pageable pageable = PageRequest.of(page, size);
+        return employeeService.findAll(position, department, active, pageable);
+    }
+}
