--- user-service/src/main/java/rs/raf/user_service/employee_search/EmployeeController.java
@@ -1,53 +1,0 @@
-package rs.raf.user_service.employee_search;
-
-import org.springframework.data.domain.Page;
-import org.springframework.data.domain.PageRequest;
-import org.springframework.data.domain.Pageable;
-import org.springframework.web.bind.annotation.*;
-
-@RestController
-@RequestMapping("/api/admin/employees")
-public class EmployeeController {
-
-    private final EmployeeService employeeService;
-
-    public EmployeeController(EmployeeService employeeService) {
-        this.employeeService = employeeService;
-    }
-/*
-    @GetMapping("/{id}")
-    public Employee getEmployeeById(@PathVariable Long id) {
-        return employeeService.findById(id);
-    }
-
-    @GetMapping
-    public Page<Employee> getAllEmployees(
-            @RequestParam(required = false) String position,
-            @RequestParam(required = false) String department,
-            @RequestParam(required = false) Boolean active,
-            @RequestParam(defaultValue = "0") int page,
-            @RequestParam(defaultValue = "10") int size) {
-
-        Pageable pageable = PageRequest.of(page, size);
-        return employeeService.findAll(position, department, active, pageable);
-    }
-
- */
-
-    @GetMapping("/{id}")
-    public EmployeeDTO getEmployeeById(@PathVariable Long id) {
-        return employeeService.findById(id);
-    }
-
-    @GetMapping
-    public Page<EmployeeDTO> getAllEmployees(
-            @RequestParam(required = false) String position,
-            @RequestParam(required = false) String department,
-            @RequestParam(required = false) Boolean active,
-            @RequestParam(defaultValue = "0") int page,
-            @RequestParam(defaultValue = "10") int size) {
-
-        Pageable pageable = PageRequest.of(page, size);
-        return employeeService.findAll(position, department, active, pageable);
-    }
-}
