+++ user-service/src/main/java/rs/raf/user_service/entity/Client.java
@@ -1,6 +1,7 @@
 package rs.raf.user_service.entity;
 
-import javax.persistence.*;
+import javax.persistence.DiscriminatorValue;
+import javax.persistence.Entity;
 
 @Entity
 @DiscriminatorValue("CLT")
